requirements.txt
Flask==0.10.1
Flask-SQLAlchemy==2.1
Jinja2==2.8
MarkupSafe==0.23
Pillow==3.2.0
SQLAlchemy==1.0.13
Werkzeug==0.11.9
argparse==1.2.1
beautifulsoup4==4.4.1
configobj==5.0.6
itsdangerous==0.24
langdetect==1.0.6
numpy==1.11.0
requests==2.10.0
six==1.10.0
wsgiref==0.1.2
